python_ai/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
numpy==1.26.4
python-dotenv==1.0.1
